requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
Django==4.2.11
django-allauth==0.61.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
idna==3.6
oauthlib==3.2.2
packaging==23.1
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==1.7.1
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.4.0
sqlparse==0.4.4
typing-extensions==4.7.1
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.5.0
